move most text label-related functions to a separate file which gets 'mixed in' to the Points style
